// src/apis/types.service.ts
export interface IImageInfoList {
  title?: string;
  source?: string;
  link?: string;
}

export interface IVisualMatches {
  position: string;
  title: string;
  link: string;
  thumbnail: string;
  source: string;
  source_icon: string;
}